Refactored Seismic Portal API:
- Added a utility module to handle common tasks
- Added a UUID to ensure uniqueness in earthquake entries. Args:
    - `region` (str): The region where the earthquake occurred
    - `timestamp` (str): The timestamp of the earthquake
    - `magnitude` (float): The magnitude of the earthquake
- Removed time_interval
