+++ services/earthquake_producer/src/seismic_portal_api/earthquake.py
@@ -13,3 +13,4 @@ class Earthquake(BaseModel):
     depth: float
     latitude: float
     longitude: float
+    uuid: str
